Object를 전부 ObjectManager에서 관리하도록 수정 중  (LinkedList) . GameObject가 Tag 가지도록 수정해야하고, Camera와 Light전부 GameObjectManager에서 가지고 있도록 수정하자.

+++ Graphics/Graphics/Object.cpp
@@ -295,7 +295,7 @@ GameObject::GameObject(const char* path, bool objFile) {
 }
 
 void Transform::SetPosition(float x, float y, float z) {
-	int dx, dy, dz;
+	float dx, dy, dz;
 	dx = x - position.x;
 	dy = y - position.y;
 	dz = z - position.z;
